+++ core/profiles/topic.py
@@ -12,6 +12,7 @@ import sqlite3
 import time
 
 from config.loader import get_config
+from core.schema import ensure_schema_version
 from core.types import TopicProfile
 
 # ---------------------------------------------------------------------------
@@ -32,6 +33,7 @@ class TopicProfileManager:
         self._conn = sqlite3.connect(db_path, check_same_thread=False)
         self._conn.row_factory = sqlite3.Row
         self._create_tables()
+        ensure_schema_version(self._conn)
 
     def _create_tables(self) -> None:
         self._conn.execute("""